requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
fpdf2==2.7.9
python-dotenv==1.0.1
google-generativeai
google
